// src/types/datatype.ts
export enum Datatype {
  Int = 'int',
  SmallInt = 'smallint',
  BigInt = 'bigint',
  Float = 'float',
  Double = 'double',
  Char = 'char',
  Varchar = 'varchar',
  Text = 'text',
  Date = 'date',
  Time = 'time',
  Timestamp = 'timestamp',
  Boolean = 'boolean',
  Bit = 'bit',
  Varbit = 'varbit',
  Json = 'json',
  Uuid = 'uuid',
  //Xml = 'xml',
  //Decimal = 'decimal',
  Serial = 'serial',
  SmallSerial = 'smallserial',
  BigSerial = 'bigserial',
}